Delete old code and rename
