+++ packages/client/src/components/Map/index.tsx
@@ -5,7 +5,7 @@ import MapComponent, { Props } from './Component'
 
 const apiKey = process.env.REACT_APP_GOOGLE_MAPS_KEY
 
-const MapNew = (props: Props) => {
+const Map = (props: Props) => {
   if (!apiKey) {
     return <MapPlaceholder text="Hmm could not load map - missing API key" />
   }
@@ -25,4 +25,4 @@ const MapNew = (props: Props) => {
   return <Wrapper apiKey={apiKey} libraries={['geometry']} render={render} />
 }
 
-export default MapNew
+export default Map
